This implements: https://github.com/ooni/api/issues/52

On some test cases using `domain` instead of `input` for filtering the target domain name we get a 100x performance gain.

See:

**Before**:
```
EXPLAIN ANALYZE
SELECT measurement.input_no AS m_input_no, measurement.measurement_start_time AS measurement_start_time, measurement.msm_no AS msm_no, measurement.report_no AS m_report_no, coalesce(measurement.anomaly, false) AS anomaly, coalesce(measurement.confirmed, false) AS confirmed, coalesce(measurement.msm_failure, false) AS msm_failure, measurement.exc AS exc, measurement.residual_no AS residual_no, report.report_id AS report_id, report.probe_cc AS probe_cc, report.probe_asn AS probe_asn, report.test_name AS test_name, report.report_no AS report_no, coalesce(input.input, NULL) AS input
FROM measurement JOIN report ON report.report_no = measurement.report_no JOIN input ON measurement.input_no = input.input_no
WHERE input.input LIKE '%%google.com%%' AND report.test_name = 'web_connectivity' ORDER BY test_start_time desc
 LIMIT 100 OFFSET 0;


Limit  (cost=6008720.69..6008720.94 rows=100 width=181) (actual time=482516.887..482517.126 rows=100 loops=1)
  ->  Sort  (cost=6008720.69..6008727.23 rows=2613 width=181) (actual time=482516.885..482516.963 rows=100 loops=1)
        Sort Key: report.test_start_time DESC
        Sort Method: top-N heapsort  Memory: 51kB
        ->  Nested Loop  (cost=1.00..6008620.83 rows=2613 width=181) (actual time=0.759..476700.235 rows=3008623 loops=1)
              ->  Nested Loop  (cost=0.57..5958840.44 rows=27056 width=73) (actual time=0.742..434003.226 rows=3235992 loops=1)
                    ->  Seq Scan on input  (cost=0.00..30516.26 rows=157 width=25) (actual time=0.000..553.824 rows=4202 loops=1)
                          Filter: (input ~~ '%%google.com%%'::text)
                          Rows Removed by Filter: 1560758
                    ->  Index Scan using measurement_v2_input_no_idx on measurement  (cost=0.57..37278.54 rows=48149 width=52) (actual time=0.139..101.491 rows=770 loops=4202)
                          Index Cond: (input_no = input.input_no)
              ->  Index Scan using report_pkey on report  (cost=0.43..1.83 rows=1 width=97) (actual time=0.008..0.009 rows=1 loops=3235992)
                    Index Cond: (report_no = measurement.report_no)
                    Filter: (test_name = 'web_connectivity'::ootest)
                    Rows Removed by Filter: 0
Planning time: 0.289 ms
Execution time: 482517.241 ms

```

**After**:
```
EXPLAIN ANALYZE
SELECT measurement.input_no AS m_input_no, measurement.measurement_start_time AS measurement_start_time, measurement.msm_no AS msm_no, measurement.report_no AS m_report_no, coalesce(measurement.anomaly, false) AS anomaly, coalesce(measurement.confirmed, false) AS confirmed, coalesce(measurement.msm_failure, false) AS msm_failure, measurement.exc AS exc, measurement.residual_no AS residual_no, report.report_id AS report_id, report.probe_cc AS probe_cc, report.probe_asn AS probe_asn, report.test_name AS test_name, report.report_no AS report_no, coalesce(input.input, NULL) AS input
FROM measurement JOIN report ON report.report_no = measurement.report_no JOIN input ON measurement.input_no = input.input_no
WHERE input.input SIMILAR TO '(https://|http://)google.com%%' AND report.test_name = 'web_connectivity' ORDER BY test_start_time desc
 LIMIT 100 OFFSET 0

Limit  (cost=6008720.69..6008720.94 rows=100 width=181) (actual time=2275.895..2276.143 rows=100 loops=1)
  ->  Sort  (cost=6008720.69..6008727.23 rows=2613 width=181) (actual time=2275.892..2275.983 rows=100 loops=1)
        Sort Key: report.test_start_time DESC
        Sort Method: top-N heapsort  Memory: 53kB
        ->  Nested Loop  (cost=1.00..6008620.83 rows=2613 width=181) (actual time=16.287..2273.799 rows=1263 loops=1)
              ->  Nested Loop  (cost=0.57..5958840.44 rows=27056 width=73) (actual time=16.270..2187.494 rows=13452 loops=1)
                    ->  Seq Scan on input  (cost=0.00..30516.26 rows=157 width=25) (actual time=16.241..2131.136 rows=154 loops=1)
                          Filter: (input ~ '^(?:(?:https://|http://)google\.com.*.*)$'::text)
                          Rows Removed by Filter: 1564806
                    ->  Index Scan using measurement_v2_input_no_idx on measurement  (cost=0.57..37278.54 rows=48149 width=52) (actual time=0.010..0.210 rows=87 loops=154)
                          Index Cond: (input_no = input.input_no)
              ->  Index Scan using report_pkey on report  (cost=0.43..1.83 rows=1 width=97) (actual time=0.004..0.005 rows=0 loops=13452)
                    Index Cond: (report_no = measurement.report_no)
                    Filter: (test_name = 'web_connectivity'::ootest)
                    Rows Removed by Filter: 1
Planning time: 0.793 ms
Execution time: 2276.324 ms
```